+++ src/useMedia/index.ts
@@ -1,1 +1,1 @@
-export * from './media'
+export * from './useMedia'
