+++ src/useSubscription/index.ts
@@ -1,2 +1,2 @@
-export * from './createActions'
-export * from './subscribe'
+export { createActions } from './utilities/actions'
+export * from './useSubscription'
