+++ src/useSubscription/models/subscription.ts
@@ -1,7 +1,8 @@
 /* eslint-disable max-classes-per-file */
 import { ref, Ref, watch } from 'vue'
+import { Action, ActionResponse } from '../types/action'
+import { SubscriptionOptions } from '../types/subscription'
 import Channel from './channel'
-import { Action, ActionResponse, SubscriptionOptions } from './types'
 
 class SubscriptionIdManager {
   private static id: number = 0
