+++ src/useSubscription/types/action.ts
@@ -1,0 +1,9 @@
+import { MaybeRefs, OnlyRequired } from './utilities'
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type Action = (...args: any[]) => any
+export type ActionArguments<T extends Action> = MaybeRefs<Parameters<T>>
+export type ActionResponse<T extends Action> = Awaited<ReturnType<T>>
+
+export type ChannelSignature = `${number}-${string}`
+export type ActionParamsRequired<T extends Action> = OnlyRequired<Parameters<T>>
