+++ src/useSubscription/types/utilities.ts
@@ -1,0 +1,10 @@
+import { Ref, UnwrapRef } from 'vue'
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type ToPossibleRefs<T> = {
+  [K in keyof T]: T[K] | Ref<UnwrapRef<T[K]>>
+}
+
+export type MaybeRefs<T extends any[]> = ToPossibleRefs<T> | Ref<ToPossibleRefs<T>>
+
+export type OnlyRequired<T extends any[], U extends any[] = []> = Partial<T> extends T ? U : T extends [infer F, ...infer R] ? OnlyRequired<R, [...U, F]> : U
