+++ src/useSubscription/utilities/reactivity.ts
@@ -1,0 +1,23 @@
+import { isReactive, isRef, unref, WatchSource } from 'vue'
+import { Action, ActionArguments } from '../types/action'
+
+export function unrefArgs<T extends Action>(args: ActionArguments<T>): Parameters<T> {
+  const argsUnref = unref(args) as Parameters<T>
+
+  return argsUnref.map(unref) as Parameters<T>
+}
+
+export function watchableArgs<T extends Action>(args: ActionArguments<T>): WatchSource | WatchSource[] | null {
+  if (isRef(args) || isReactive(args)) {
+    // can't quite figure out the types here. But the tests around reactive arguments pass so I believe this is correct
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return args as any
+  }
+
+  if (Array.isArray(args)) {
+    return args.filter(arg => isRef(arg) || isReactive(arg))
+  }
+
+  return null
+}
+
